[PORT] TGUI Stack Zebra Mode (#5891)
## About The Pull Request
* Ports tgstation/tgstation#79972

One of the many PRs that solves merge conflicts with the React port.
(Yes, I'm still working on it. I'm going slowly, but Inferno is serving
us well for the moment, so I'm not in too much of a rush.)

Also just looks nice.

## Changelog

:cl:MichiRecRoom
add: (Ported from tgstation - ZeWaka) TGUI Stack elements can now be
zebra-styled (alternating colors for contrast)
/:cl:

Co-authored-by: ZeWaka <zewakagamer@gmail.com>
Co-authored-by: oranges <email@oranges.net.nz>