Murderbone Cape & Hat (#5653)
## About The Pull Request
Reverts: https://github.com/Monkestation/Monkestation2.0/pull/5387

Adds a Corporate Sponsorship for Evil bundle to the uplink (basically an
excuse to murderbone)
- 20 TC purchase, has to be the first purchase, locks the uplink
- Once putting on the provided cloak it alerts ghost and admins that you
are allowed to murderbone.
- The hat gives buffs: space proofing, fearless, and immunity to pain.
- The hat gives you a red aura once worn.
- The cloak gets stuck to you once put on and can't be removed.
- You can't remove the hat yourself but if someone else does it gibs you
and then explodes.

## Why It's Good For The Game
An invisible murderbone pass makes for an ahelp generating machine. This
lets people know the person attacking them has the pass.

## Changelog

:cl:
add: Added the Corporate Sponsorship for Evil bundle to the uplink. 20
TC purchase for traitors as a pass to murderbone
del: Reverted the License to Murderbone to License to Plasmaflood
fix: adjusted the fake mustache sprite to look better
/:cl:

---------

Co-authored-by: SirNightKnight <=>