Make the Recolorable Turtleneck available to all (though donators still get the item for free) (#5946)
## About The Pull Request
This PR adds the Recolorable Turtleneck loadout item to the store,
allowing all to enjoy its benefits.

This was previously restricted to donators, as it was a donator item
intended for TheSpecialSnowflake - however, TheSpecialSnowflake gave
permission to make this available to all.
<https://discord.com/channels/748354466335686736/808983978459004968/1350629297790128209>

To avoid donators losing the item (since they've already had the item),
donators still get the item for free. The way I handled this ends up
allowing donators to have TWO recolorable turtlenecks in their inventory
(and as such, I've marked one as "Donator" so people aren't confused).
However, this shouldn't be an issue - it'll just look a little goofy.

## Why It's Good For The Game
More colorable clothes for all!

## Changelog

:cl:MichiRecRoom
add: The Recolorable Turtleneck loadout item is now available to all!
You can find it in the "Misc. Under" section of the store for 3000
Monkecoin. Donators are unaffected by this change and still receive a
Recolorable Turtleneck for free (though you can purchase the item
anyways if you want two Recolorable Turtlenecks in your inventory).
/:cl: