3D printed guns. (#4814)
## About The Pull Request
The FSS-550, a version of the WT-550 that can be printed from an
autolathe with a disk. Is worse than the WT-550.
(I need advice on balancing and stats and whatnot.)
FSS-550 compared to WT-550:
-0.9 damage modifier.
-Takes 4 seconds of standing still to take a magazine in or out.

The FSS guns and ammo can be printed in an autolathe with a design disk.
Ways to get the design disk:
-Traitors can buy it for 5 tc.
-The black market has a 10% chance to have it available.
-An emagged supply console can buy it. (If you have an emag and cargo
access you could already just buy better guns.)
-Rare spawn in maintenance.

The black market also has a 20% choice to sell actual FSS guns. (Maybe
should remove this?)

## Why It's Good For The Game
You can't print guns despite that being a very much real thing. So
making a possible way to print guns could be fun if it's locked behind
some stuff.

F*** S*** Sec!

## Changelog
:cl:
add: Adds the FSS-550 3D printed gun from the autolathe. It requires a
special design disk to make.
/:cl: