Chatbox Websocket client (#5744)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds a websocket client to the chatbox, that redirects all
incoming payloads from the server, to the specified server, which allows
for external integrations with the game. All of which is parsable JSON.

Some ideas include RGB feedback, triggering an event in VTubing
software, or automating a physical device based on what's sent to the
chat. (For example, a haptic suit, and have a regex for "in the right
leg" or any body part)

Works on 515 and 516. (I didn't realize IE11 had websockets until I went
to go and make sure my code to disable it worked, then had to double
check myself because it was connecting to my websocket server, and
working. so. LOL)

Here's a barebones snippet, not feature complete but it has enough where
you can start with phrase
```js
const ws = require("ws");
const phrases = require("./phrases");

const LISTENPORT = 8094;
const wss = new ws.Server({ port: LISTENPORT });

wss.on("listening", () => console.log(`Listening on ${LISTENPORT}`));

wss.on("connection", (socket) => {
  console.log("Client connected!");

  socket.on("message", async (data) => {
    try {
      const messageData = JSON.parse(data.toString("utf-8"));
      if (messageData.type !== "chat/message") return;
      const parsedPayload = JSON.parse(messageData.payload);

      if (!parsedPayload?.content?.html) return;

      const message = parsedPayload.content.html.toLowerCase();
      const matchedPhrase = phrases.phrases.find((phrase) =>
        message.match(phrase.regex)
      );

      if (matchedPhrase) {
        // Catched phrase could be something like "stabs you in the chest with the kitchen knife"
        // /(userdanger'>(?:.*you.*in the.*with the))/gi
        
        // example, setup a websocket client to connect to vNyan and send the command "modelBleed"
        if (matchedPhrase.type === "vnyanAction")
          vnyanWs.send(JSON.stringify(matchedPhrase.command))
      }
    } catch (e) {
      console.error("Message Processing Error:", e);
    }
  });

  socket.on("close", () => console.log("Client disconnected."));
});
```

```js
modules.export = {
  phrases: [
      {
        regex: /(userdanger'>(?:.*you.*in the.*with the))/gi,
        type: "vnyanAction",
        command: "modelBleed"
    },
  ]
}
```

## Why It's Good For The Game

QoL for people who want things outside of the game to respond to in-game
actions. See above examples.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Chatbox websocket feature
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Lucy <lucy@absolucy.moe>