Move Research techweb to serveral files rather than one big one (#5916)
## About The Pull Request
Moves as much research as possible from `all nodes` to

- Alien_nodes
- atmos_nodes
- Bepis_nodes
- biology_nodes
- circuit_nodes
- cyborg_nodes
- engi_nodes
- mech_nodes
- medbay_nodes
- mining_nodes
- modsuit_nodes
- research_nodes
- security_nodes
- service_nodes
- surgery_nodes
- syndicate_nodes
## Why It's Good For The Game
### 2541
It's a 2541 long line file, with some organization in it.
The reason we have the file is the reason to my understanding is the
same reason as we have /code. To keep code easily portable into monke
from TG.

The catch? TG doesn't use All_nodes anymore if they ever did. Instead
they use the files listed above that were created.

Our research arguments and overall tech tree is so very different from
TG now. But this helps brings it a little in line on the code side.
Everything was placed close as reasonably possible to its original
locations. Differences still occur of course because of how far we and
TG research has branched, but its better than nothing.

The main catch to this PR is it will, without a doubt, cause merge
conflicts.
## Changelog
:cl:
code: Research Tech Tree was split amongst several files instead of one
very long one.
/:cl: