[Bounty] Head of __ (#5950)
## About The Pull Request
Adds Head of Science and Head of Command as alternative titles for
Research director and Captain at bounty request.
## Why It's Good For The Game
All other heads of staff members have alternative titles with the format
head of _. Finish the trend. Finish what is not compelete
## Changelog
:cl:
add: Added Head of Command and Head of Science as alternative titles for
Captain and Research Director.
/:cl: