[BOUNTY] Implements atom integrity for walls, (which is needed to do) Upgraded Emitters can now tear down walls! Take Two! (#5938)
## About The Pull Request
Ports yogstation13/Yogstation#21336
Retake of #5706
This PR is a double bounty, makes it so that walls behaves the same way
as any other structures or machineries, being able to be damaged and
destroyed by normal weapons instead of very specific ways/deconstruction

Normal walls have 300 damage + armor, with a damage minimum for it to
take effect of 25
Normal walls have 600 damage + armor, with a damage minimum for it to
take effect of 75

Emitters with upgraded lasers can now be able to tear down walls, stock
parts emitters still functions the same. Below are the damage table
needed to tear a wall down

T1 laser: cannot damage walls, operates normally.
T2: 20 hits to break normal walls, r wall or stronger cannot be damaged.
T3: 10 hits to kill normal walls, r wall or stronger cannot be damaged.
T4: 5 hits to kill normal walls, r walls or strong take 10 hits to be
killed.
T5: 3 hits to kill normal, 7 to kill r walls or higher
Emagging the emitters increases the parts tier above by one (which is
how you'd get T5) at the cost of 5% damage to the emitters. This
functionality only applies if the projectile emitted by the emitters is
the default hitscan emitters shots.

Pooba said that for my first PR to go through, we'd need to implement
atom integrity, which they'd gracefully posted a bounty for.
## Why It's Good For The Game
Here is it in action:

https://github.com/user-attachments/assets/34b999b0-a7f6-4fa7-ac26-5e804e3ba5ad

Walls being able to use the integrity system means destroying them in a
way that isn't instantaneous and no longer needs to rely on tons of
snowflake code for each instance/case of something breaking a wall. It
also means they aren't totally invincible to damage, which I think could
allow for some fun interactions.

Anyhow, as for the emitter PR. i'll just copy this verbatim: You know
how the emitters looks like a heavy, industrial laser cannon that you'd
be able to see be used as an impromptu blaster? How come can it not
damage walls? Anyways, this adds a new way of breaching walls besides
thermite, tools or c-4/x-4, Albeit a pretty slow and loud one at that.
More tools for engineer's arsenal is good!
## Changelog
:cl:
add: Walls now have atom integrity, being able to be attacked and
destroyed like other machineries and structures. they're a tough nut to
crack though
 add: Emitters can now tear down walls! (once upgraded)
/:cl: