Adding some substance to the detective role + a port from TG (#5970)
## About The Pull Request

-> Make two roundstart detective slots
-> Added a second holster to the detective's closet that posses a second
detective revolver (that's just a normal .38)
-> Added an undercover I.D, same as an agent I.D without the capacities
to copy access.
-> Spyglass got sechud
-> Added Kim's aerobic gloves to the loadout
-> Added a .45 rubber ammo for the pistol
-> Port from TG : https://github.com/tgstation/tgstation/pull/89221

## Why It's Good For The Game

Often as a detective, I try to find case, to find a buddy or a partner.
Yet I find none of that, security assistants with the forensic kit
ignore me/outright refuse to tag along. I can't get any real case going
and going undercover is a nightmare.

With this, I get a buddy, a way to be undercover, a way to make a case.
Detect or DIE

## Changelog

:cl:
add: Added a detective slot and a second holster in the detective closet
+ detective spawn with a badge.
add: Added Kim's gloves in the loadout, exclusive to the detective +
spyglass got sechud.
add: .45 caliber rubbers available in the security techfab
add: Ported from TG : https://github.com/tgstation/tgstation/pull/89221
which mean the forensic scanner got a new ui
/:cl: