Lowers depot TC from 25 to 5 (#6056)
## About The Pull Request
Lowers amount of TC in the depot safe from 25 TC to 5 TC

## Why It's Good For The Game
This was after discussion with quite a few people including addust who
made the depot.

25 TC is absured and over tots starting TC for free with a fairly often
spawning space ruin. People hit up the depot nearly every chance they
get in my experience, which is a lot of chances.
Depot has plenty of reasons to visit and work with them outside of
giving them a insane amount of TC, this gives them a little bit but not
a ton instead.

## Changelog

:cl:
balance: Depot Safe TC lowered from 25 to 5
/:cl: