Makes the heroic laser musket rarer (#6036)
## About The Pull Request
lowers the heroic laser musket chance of appearing in black market from
40% to 5% with a max of 1

## Why It's Good For The Game
The laser heroic musket got massively buffed to being able to recharge
while moving and shooting 3 25 damage laser projectiles.
Before this it was a terrible weapon with only a single 30 damage
projectile that took several seconds of standing still to recharge. It
was during this period that it was really derarified, although being a
rare maints book it was given to blood brothers, a high chance in the
black market, and mapped in a few places.
After its buff none of that was changed, being a extremely powerful gun
and one of the single best in the game that is fairly common and can be
obtained just about every round.

Considering all that this still leaves the gun fairly common and
arguably still mroe common than it should be. Might just be due for a
damage nerf in general and then made a little less rare again. Alt
option to this PR.

## Changelog

:cl:
balance: the heroic laser musket now appears in the black market less
/:cl: