[Code Bounty] Lisp Quirk (#5905)
## About The Pull Request
Adds lisp quirk for -2 points. It is incompat with mute

replaces S with th

taken with permission from
https://github.com/Monkestation/Monkestation2.0/pull/4907

## Why It's Good For The Game
Accents are a bit of a pain in the ass to mantain 24/7 although they are
great for playing a character, this takes away a bit of that pain,
adding a nice little accent for people to RP with that also leaves them
understandable.

I possibly want to build on this quirk in the future making it neutral,
and giving it a drop down with some other MINOR accents (nothing like
fluffy tongue)

## Changelog

:cl:
add: Lisp Quirk
/:cl: