Fix a lot of improperly mapped cameras (#6150)
## About The Pull Request

This mass replaces all `/obj/machinery/camera/directional/[DIR]` that
didn't manually set name / tag / networks with
`/obj/machinery/camera/autoname/directional/[DIR]` - which generates a
name + tag on them during initialize, allowing them to, for example, be
seen with a camera console.

## Why It's Good For The Game

it's nice to be able to actually see shit with a camera console

## Changelog
:cl:
map: Fixed a lot of cameras that were improperly mapped in, resulting in
them not showing up on camera consoles.
/:cl: