Fix inputs on BYOND 516 getting cut off (tgui payload chunking) (#6088)
## About The Pull Request

Ports https://github.com/tgstation/tgstation/pull/90295

> Dream Seeker will not send topic calls greater than 2kb in size. There
are cases where tgui will attempt to send ui_act payloads larger than
this, such as writing on paper. This PR takes payloads that would be
larger than 2kb, splits them into payloads that would be roughly 1kb
(after URL encoding), and sends them to the server in sequence. To
prevent abuse and/or topic spam, a config option has been added to put a
limit on the number of chunks for which the server will accept a
payload, defaulting to 10.

## Why It's Good For The Game

> Fixes (tg issue 90056), along with several other things that were
affected by the change to WebView2 in 516.

## Changelog
:cl: Absolucy, Y0SH1M4S73R
code: Any tgui message that would be too big to send to the server is
now split into chunks and sent in sequence. This fixes several issues,
such as...
fix: It is once again possible to save large amounts of text on paper at
once.
/:cl:

---------

Co-authored-by: Y0SH1M4S73R <legoboyo@earthlink.net>
Co-authored-by: Y0SH1M4S73R <y0sh1m4s73r@gmail.com>