Martial arts properly dodge/counter Kindle (#5978)
## About The Pull Request

this makes it so cogger Kindle is also dodged/countered by martial arts,
like the blood cult stun and mansus grasp

the counter is a bit different tho - as it's being channeled through a
physical object, instead of hitting the attacker with their own spell,
they'll be forced to drop the slab instead.

## Why It's Good For The Game

consistency.

## Changelog
:cl:
balance: Kindle is now dodged/countered by martial arts, consistent with
the mansus grasp and blood cult stun.
/:cl: