Adds the "Unborgable" quirk (#6154)
## About The Pull Request

This adds a new neutral quirk, Unborgable.

It does exactly what it says on the tin - your brain can never be used
for a cyborg or MMI.

An unborgable person going thru a malf AI's borg factory will just be
converted into an infected IPC instead, regardless of the setting of the
factory.

## Changelog
:cl:
add: Added a new neutral quirk, "Unborgable". It's exactly what it says
on the tin, and prevents your brain from being used for a cyborg or AI,
for better or worse.
/:cl: