Refactor armaments and imports to use `DmIcon` (#6216)
## About The Pull Request

This makes it so armaments + imports just fully use `DmIcon` for UI
rendering.

This saves a second or so of init time, due to not needing to do a bunch
of `icon2base64(getFlatIcon(...))` in init, and also heavily reduces the
amount of data sent to users when opening the UIs.

## Why It's Good For The Game

Reduced init time, less garbage being sent in `ui_data`

## Changelog
:cl:
refactor: The cargo imports console and assault op armaments uplink now
use a more efficient method of displaying icons.
/:cl: