Allow antag tokens to be used for revenants (if configured) (#6207)
## About The Pull Request

This adds code for properly spawning a revenant if someone uses a token
to become one.

THIS DOES NOT ADD THEM TO THE ANTAG TOKEN CONFIG, THAT NEEDS TO BE DONE
SERVER-SIDE, THIS JUST MAKES IT SO IT WORKS AS EXPECTED WHEN THEY ARE


![image](https://github.com/user-attachments/assets/d737237b-bd1e-42c6-b4f1-9bfac89e5a70)

## Changelog

No changelog, as this requires server-side config changes, and it will
likely confuse players as a result.