What is 52.3 research points? (Research point generation) (#6026)
## About The Pull Request
Accounted for nanites artsci and experiments generation so far
Converts all research, and research requirements to a much lesser, and
more comprehendible number.

- Passive research 52.3 per second -> 1 per second
- tier 1 research 2500 points -> 40 points
- research nanites 0.5 -> 0.01

To gives some numbers that qauntify what these changes mean. Research
current takes 174 minutes. With this change It will take 154 . All
values are effected from being 0.956x slower to up to 1.156x faster
roughly

These are more side effects than the actual intention of the PR.
## Why It's Good For The Game
52.3 points per second communicates, nothing. It's an arbitrary rubric
number that the rest of the research works around. You cant gauge how
long a research will take off hand, (A 2.5k research will take 47
seconds). It serves a detriment to communication and understanding. So
my proposal is to change it to a number that communicates more. 1 point
per second.

This way when you see a research with 200 points to unlock it. It
communicates a solid peace of information. Namely the time, this will
take 3 minutes and 20 seconds to unlock via waiting. So you should go do
a discount instead
## Changelog
:cl:
balance: Research now generates at 1 point per second, all other metrics
based on this number follow suite.
/:cl: