Positronic Reboot Surgery fix (#6038)
## About The Pull Request

The behavior of the reboot actually passes the "failed" text when the
surgery is successful. This restructures it so it actually works and
tells you why the reboot doesn't work (if applicable)

## Why It's Good For The Game

Readability on a lot of IPC repair stuff is bad. One step at a time.

## Changelog
:cl:
fix: IPC reboot surgery now displays the correct error/complete
messages.
balance: IPC reboot surgery now requires 5 iron in the step, not 15.
/:cl:

---------

Co-authored-by: Ziiro <thechaosofzero@gmail.com>
Co-authored-by: Lucy <lucy@absolucy.moe>