[BOUNTY] SCRAP VEHICLES! Locker mechs, ambulance and scrap tank! (#6165)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds scrap vehicles. vehicles (mechs in disguise) that can be made from
the build menu. They are:
1. Locker mech: a makeshift version of the ripley, featuring essentially
zero hp, inferior storage space and a limited module selection. It's
ported from old monkestation [(
#329)](https://github.com/Monkestation/MonkeStation/pull/329). I
refactored it to be less shit with it's copied pasted code though.

2. Portaporty ambulance: What it says on the tin, it's a shitty
ambulance that can only really used as a stretcher. oh and it comes with
a 1-tile HoNkBlAsTeR as a harm alarm to knock over folks hindering first
aid.

3. Trash tank: A Literal trash tank. It's only marginally better
protected than the first two scrap vehicles. Can actually protect
against stuff better than a crowbar with further upgrade though. it has
a pipegun and peashooter turret gun. the pipegun is essentially just a
pipegun that you can use in the tank. the peashooter shoots a burst of
improvised shells pellets with no damage falloff


![image](https://github.com/user-attachments/assets/8032ea64-2c3d-443a-ad86-7bf29496c9a1)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Gives assistants with more time than they should have something fun to
do to keep them distracted from causing havoc. great makeshift weapons
that has some teeth while requiring a lot of effort to make.

Also, the sight of a platoon of tanks roaming the hallways is very
amusing.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Adds scrap vehicles (Locker mechs, Ambulance and trash tank!)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->