Updates Piecakes Piecakery and touched up my admin office because I am a noncommital bitch (#6359)
## About The Pull Request
Touched up Piecakes admin area as it was begging to be touched up as it
went unmodified for a while, per request. I also fixed up my office
(again) because I have a brainworm that tells me what to do and today it
wanted a better office.

## Why It's Good For The Game
Its not

## Changelog
:cl:
map: Modifies the centcom z level and modifies 2 admin offices
/:cl: