Cortical Borer Egg fixes (#6659)
## About The Pull Request
Hive lord self reproduction ability is now functional. Fixes an old bug.
Fixes:#4334
Chemical costs for producing a egg is now subtracted after the egg is
produced.
## Why It's Good For The Game
So it checked for a host when the proc itself has interal checks for
hosts. And this made self reproduction nonfunctional. This allows for
specialized Hivelords to self-reproduce as intended

And the chemical cost has been moved to reflect only when an egg is to
be reproduced, not beforehand. Stopping chemical loss for nothing.
## Changelog
:cl:
fix: Hivelords Borers now can self reproduce as intended.
fix: Stopped an edge case where borers would lose chemicals but produce
no eggs
/:cl: