Quick fix of outfit permissions. Whoops in master branch but we ball. (#6692)
## About The Pull Request
Whoopie, Forcepush to master
Fixes a minor oversight in outfits
## Why It's Good For The Game
Fixes a minor oversight in outfits


## Changelog
Fixes a minor oversight in outfits


:cl:
fix: Quick fix of admin outfit permissions


<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->