NTSL compile errors and save changes (#6690)
## About The Pull Request

This PR does three things:
- It adds a log message whenever the NTSL traffic console save your
code.
- It breaks off the max character check from the invalid character check
and into its own if statement, displaying how many characters the saved
code has and the max character limit in the log.
- It lowers the max character limit from 20,000 characters to 15,000
characters.

## Why It's Good For The Game

For code saving, this is helpful for larger scripts, since the UI is
slow to save them. If a script is compiled before it is saved, it either
throws an error because it thinks you tried to compile nothing or it
compiles whatever it had saved previously. With the console log now
telling you when it is saved, you no longer have to wait for the button
press sound to know that it saved your code.

For the error checks, before this PR, one method checked for both
invalid characters and if the script was over the max character limit.
This is bad because if you had a script that was too big, rather than
tell you that, it would say that you had an invalid character. With the
checks now separated, along with it showing your script's character
count and the max character count, it is now easier to troubleshoot NTSL
compile errors.

For lowering the max character count, TGUI does not seem capable of
compiling scripts anywhere near 20,000 characters, with it starting to
break down around ~16,000-15,000 characters. With the max character
count lowered to 15,000, this will still allow for large scripts while
also lessening the affects of TGUI having difficulty with them. Even
around 15,000 characters, TGUI is slow to save NTSL code as it takes
around 5-10 seconds for it to trigger the code that saves your NTSL
script.

##Testing

https://www.youtube.com/watch?v=qFo-1nuYZwE

## Changelog

:cl:
qol: Broke NTSL invalid and max character checks into separate checks
for easier troubleshooting
fix: Lowered NTSL max characters to 15,000 due to TGUI limitations
/:cl: