Unique Heart Effect (#6816)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Example

https://github.com/user-attachments/assets/b1667702-edd2-4ba7-9d63-a97080dcf74c



Just a heart emitter effect to add to the unique effects that can be
rolled in loot boxes and stuff. Uses the heart originally from the
valentines day event.

## Why It's Good For The Game

Just another unique added. Uses the code for the fire unique and icon
from the valentines day event. Nothing major we don't already have adds
more to the game and we already know it works.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:Siro
add: Heart emitter unique effect
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Siro <yamapeachsan@gmail.com>