Slasher antag tweaks (#6714)
## About The Pull Request
Slight balance tweak to the Slasher's Jaunt.
## Why It's Good For The Game
The previous jaunt was a long cooldown, for hardly any actual movement.
The new jaunt change increases the jaunt time by double and also removes
cuffs from the slasher. This justifies the two minute cooldown by
turning the jaunt into a proper escape spell for when you are surrounded
or cuffed and dragged by a single sec officer.
## Changelog
:cl:
balance: Buff to the Slasher's jaunt.
/:cl: