Adds the ability for an AI to remotely charge an APC with an upgrade disk #86470 (#6562)
## From #86470
https://github.com/tgstation/tgstation/pull/86470 orginal pr

To-do
Makes it so incapacitated ais cant use it (Orginal proc doesnt work)
## About The Pull Request
Adds an upgrade disk that gives the ability to remotely charge an APC
with the AIs backup power (the one that runs down when there is no power
in an APC). This tech disk is under advanced AI tech

Also added the ability to make new disks with the new object type
/obj/item/aiupgrade, and made it so not every AI module is malf, now
only subtypes of /datum/ai_module/malf show up on malf upgrade screens
## Why It's Good For The Game
Adds the ability for an AI to assist crewmembers who do not have power,
which is a pretty interesting interaction. Does chem not have power and
is trying to make some healing chems? you can provide power for them for
a bit, but they better hurry!
## Changelog
:cl:
add: Added the remote power AI disk
code: made it possible to easily add new AI upgrade disks
/:cl:

---------

Co-authored-by: Gw0sty <83688318+Gw0sty@users.noreply.github.com>
Co-authored-by: Lucy <lucy@absolucy.moe>