Syndicate Mini-Kits and other additions (#6874)
## About The Pull Request

- Adds a new 10 TC Random Kit Bundle in the traitor's uplink called
"Syndicate Mini-Kits" with 5 random loadouts and a new sprite for the
box.

- Adds a new type of Paper that will incinerate after 20 seconds if its
read when armed, This is used as a guide note in the Mini-Kit

- Adds new Syndicate Pill Canisters that have their own sprite and a new
Poisonous Pill Canister that contain 2 cyanide pills, 2 toxin pills, 2
spewium pills, and 1 amanitin pill.

- Adds a new "can_be_folded" variable thats set to true by default.
Setting it to false will make it so the plane cant be folded into an
airplane and removes the associated examine text.

## Why It's Good For The Game
These Syndicate Mini-Kits are meant for new traitors to understand and
help them with the rough-basics of being a traitor with each different
kit teaching a separate "lesson", while also not spending all of the
player's TC in one purchase.

The Amateur Assassin Loadout is used to help people learn to kill/kidnap
discretely:
- Poisonous Pill Canister
- Sleepy Pen
- Nitrile Gloves
- Syndie Soap
- Switchblade
- Cuffs
- Sticky Tape (Used to tape people's mouths shut)
- Blindfold
- Thermal Goggles

![image](https://github.com/user-attachments/assets/9bb7ea8a-f25b-4d4f-ba79-af615ff17792)

The Intern Infiltrator Loadout is used to help people learn how to
disguise themselves and break into places:
- Syndie-Toolbox
- Chameleon Kit
- Syndicate Encryption Key
- Agent ID
- Uplink Implant

![image](https://github.com/user-attachments/assets/e7dff114-1790-430c-be25-64954da1809a)

The Beginner Bomber Loadout helps people learn to make the most out of
their limited but powerful tools (in this case explosives):
- Syndie-Minibomb
- X-4 Charge x2
- C-4 Charge x3
- Frag Grenade x2

![image](https://github.com/user-attachments/assets/1c8dd639-c00e-4b5c-a4f2-1172c251e9c8)

The Starter Saboteur Loadout is used to help teach people how to mess
with different machines using the emag and be a nuisance while not
getting detected by things like the AI, through actions such as
subverting borgs:
- Emag
- Jammer
- Jaws of Life
- Balaclava
- Ai Detector
![Screenshot 2025-06-01
211851](https://github.com/user-attachments/assets/6d990d38-f24e-4263-b3c9-4e0cd4314291)

The Rookie Raider Loadout is used to help people properly fight in
combat while going loud, while teaching them things on what to prepare
for such as slipping and stamina:
- Wespe Pistol
- Weapons Authorization kit
- Sol pistol magazine
- Stimulant Medipen
- NoSlip Chameleon Shoes
- Armor Vest
- Sol .35 Short Lethal Ammo Box

![image](https://github.com/user-attachments/assets/aa8b0c74-2689-46ad-82ee-065e576c5906)

## Changelog

:cl:
add: Added Syndicate Mini-Kits, a 10 TC traitor uplink bundle with 5
random loadouts and a unique box sprite.
add: Added 'Self-Incinerating Paper' which incinerates after 20 seconds
of being read if its armed
add: Added Syndicate Pill Canisters with new sprites, including a
poisonous variant with cyanide, spewium, amanitin and toxin pills.
add: Added 'can_be_folded' variable for paper to control if they can be
folded.
/:cl:

---------

Co-authored-by: tgstation-ci[bot] <179393467+tgstation-ci[bot]@users.noreply.github.com>