Makes both detective revolvers detective revolvers. (#6890)
## Why It's Good For The Game

I get why the person that added a second detective slot wanted only one
copy of an objective item, but honestly two copies is better than the
current solution. Every tot that wants to engage in thievery RP with the
det basically has to flip a coin on whether they're carrying THE
detective's revolver or just a detective revolver. Also having two
identical items in literally every way except for the name and that
mattering for an objective hilarious.

## Changelog



:cl:
qol: Both detective's revolvers are now actual Detective's Revolvers.
/:cl: