Firing the BSA now completes the station goal (#7003)
## About The Pull Request

this makes it so _firing_ the BSA is what completes the station goal,
which seems to be what people expect it to be.

i've updated the report text to make it clear that it needs to be test
fired.

## Why It's Good For The Game

it seems this is how people expect it to behave, and this way does make
more sense.

## Changelog
:cl:
qol: The BSA station goal is now completed by FIRING the BSA, rather
than having it just exist + be powered.
/:cl: