ports the disruptor pistol, wraith module, and infiltrator suit hides your identity slightly better (#7038)
## About The Pull Request

Ports the following PRs:
- https://github.com/tgstation/tgstation/pull/78128
- https://github.com/tgstation/tgstation/pull/78330
- https://github.com/tgstation/tgstation/pull/81553
- https://github.com/tgstation/tgstation/pull/84675
- https://github.com/tgstation/tgstation/pull/84952
- https://github.com/tgstation/tgstation/pull/86449
- https://github.com/tgstation/tgstation/pull/86546

## Why It's Good For The Game

<details>
<summary><h3>John Splintercell: a gun that is only good for shooting out
lights</h3></summary>

Adds a silly little tool for silly little men who either really hate
lightbulbs or want to recreate the experience of being John
Splintercell, the lightbulb-assasinating secret agent from Fork Echelon.

</details>

<details>
<summary><h3>Buffs the SC/FISHER Saboteur Handgun</h3></summary>
The concept is cool, alas it's also undermined by how much of a joke
it's right now, and the game has plenty already.
The amount of interactions it has with things is underwhelming, so you
could barely consider it a stealth tool. The duration is also quite
scarce, I pointed that out in the original PR too.

Basically, I want to make the item cooler.
</details>

<details>
<summary><h3>New Syndicate Stealth MODule: Wraith</h3></summary>

I've been playing a ton of Splinter Cell and Intravenous recently and
this random idea popped in my head.

"Wouldn't it be cool if traitors could blend in the darkness to get a
jump on their opponents?"

Also unrelated

"Wouldn't it be cool if tots had a tool to recharge their suit that
didn't involve sitting in a pod for 10 minutes?"

This PR introduces a new module to the uplink, the Wraith.

It comes with a passive and active component.

Passively it works exactly like the crew version of the cloaking module
with just a couple of differences.

1) Doesn't need to be manually activated, if you lose the cloak it's
regained after 5 seconds.

2) Lower stealth alpha value( how trasperent you are basically),
slightly less visible than the crew version, not as good the ninja
module however , I tuned it just enough so that you are more or less
undetectable in the dark.

The active component of the module lets you destroy stationary lights to
recharge your suit power, if used on handheld or borg lights it turns
them off for a minute.

**Why do we need this module when we already have the stealth implant
and the chameleon projector?**

I can think of a few reasons.

1) MODsuits were designed to be customizible, traitor suits range
between 6 to 16 TC, having to invest in a 7-8 TC item after you already
bought a suit is fairly expensive.

2) This MODule would be a better fit for ambushes, as it doesn't have
the *uncloaking* delay of its counterparts.

It is however considerably worse if you get caught, as the cloak is
disrupted on bump or damage.

3) It has better interactions with the sandbox.

Lights can go out for many reasons, maybe it’s just a power outage, or
some assistant broke it, or maybe it was anightmare.

 It leaves room for plausible deniability, adding to the paranoia.

It's also not complete invisibility, if you want to stay undetected you
need to lurk in the darkness, you might expand your domain, at the cost
of the crew eventually wising up to your shaeneningans.

Lastly, since the active component of the module uses the same proc of
the saboteur handgun, I've updated the code to be a generic proc rather
than a signal, to make it easier to reuse in the future.

Item desc provided by NecromancerAnne.

Module sprite made by Orcacora.

</details>

<details>
<summary><h3>Infiltrator suit now hides wings, antennae and obfuscates
your voice</h3></summary>

Further helps the infiltrator suit hide your identity by hiding
wings/antenna and makes it where you no longer have species-specific
"say" verbs or lisps.

</details>

## Changelog
:cl:
add: (Hatterhat) The SC/FISHER disruptor pistol, a very compact,
permanently silenced energy gun, is now stocked in Nanotrasen-accessible
black markets with a price generally somewhere between 400 and 800
credits. Aspiring users are warned that it's really bad for trying to
actually kill people. Caveat emptor.
add: (Ghommie) Added saboteur interactions with radios, pAIs, turrets,
secbots and APCs.
add: (Hatterhat) Guns now have a dry_fire_sound_volume variable,
allowing for guns to be less loud when trying to fire while empty.
fix: (Hatterhat) Closets and crates now properly count as structures for
pass flags again.
refactor: (SmArtKar) Refactored a lot of speech modifiers to use a
component instead of copied over code.
add: (EnterTheJake) The Wraith Cloaking Module is now available in the
uplink, costs 3 TC.
fix: (Xander3359) Infiltrator suit now hides moth wings/antenna.
fix: (Xander3359) hidden moth wings are no longer capable of flight.
/:cl:

---------

Co-authored-by: Hatterhat <31829017+Hatterhat@users.noreply.github.com>
Co-authored-by: Hatterhat <Hatterhat@users.noreply.github.com>
Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>
Co-authored-by: Jacquerel <hnevard@gmail.com>
Co-authored-by: EnterTheJake <102721711+EnterTheJake@users.noreply.github.com>
Co-authored-by: Xander3359 <66163761+Xander3359@users.noreply.github.com>
Co-authored-by: SmArtKar <44720187+SmArtKar@users.noreply.github.com>
Co-authored-by: FearfulFurnishing <139661819+FearfulFurnishing@users.noreply.github.com>